querio_backend/app/config.py
from dotenv import load_dotenv
import os

load_dotenv()

class Settings:
    SUPABASE_DB_HOST = os.getenv("SUPABASE_DB_HOST")
    SUPABASE_DB_PORT = int(os.getenv("SUPABASE_DB_PORT", 5432))
    SUPABASE_DB_NAME = os.getenv("SUPABASE_DB_NAME")
    SUPABASE_DB_USER = os.getenv("SUPABASE_DB_USER")
    SUPABASE_DB_PASSWORD = os.getenv("SUPABASE_DB_PASSWORD")

settings = Settings()